+++ ui/src/components/dashboard/components/DashboardEditor.vue
@@ -47,7 +47,7 @@
         </el-button>
     </div>
     <div class="w-100 p-4" v-if="currentView === views.DASHBOARD">
-        <ChartsSection :charts="charts.map(chart => chart.data)" />
+        <Sections :charts="charts.map(chart => chart.data)" />
     </div>
     <div class="main-editor" v-else>
         <div
@@ -124,7 +124,7 @@
     import {YamlUtils as YAML_UTILS} from "@kestra-io/ui-libs";
 
     import PluginDocumentation from "../../plugins/PluginDocumentation.vue";
-    import ChartsSection from "./ChartsSection.vue";
+    import Sections from "../sections/Sections.vue";
     import ValidationErrors from "../../flows/ValidationError.vue"
     import BookOpenVariant from "vue-material-design-icons/BookOpenVariant.vue";
     import ChartBar from "vue-material-design-icons/ChartBar.vue";
@@ -142,7 +142,7 @@
     import yaml from "yaml";
     import ContentSave from "vue-material-design-icons/ContentSave.vue";
     import intro from "../../../assets/docs/dashboard_home.md?raw";
-    import Markdown from "./MarkdownPanel.vue";
+    import Markdown from "../sections/MarkdownPanel.vue";
     import TimeSeries from "./charts/custom/TimeSeries.vue";
     import Bar from "./charts/custom/Bar.vue";
     import Pie from "./charts/custom/Pie.vue";
