+++ ui/src/components/dashboard/sections/MarkdownPanel.vue
