+++ ui/src/components/dashboard/sections/Sections.vue
@@ -40,12 +40,12 @@
     const route = useRoute();
     const router = useRouter();
 
-    import TimeSeries from "./charts/custom/TimeSeries.vue";
-    import Bar from "./charts/custom/Bar.vue";
+    import TimeSeries from "./TimeSeries.vue";
+    import Bar from "./Bar.vue";
     import Markdown from "./MarkdownPanel.vue";
-    import Table from "./tables/custom/Table.vue";
-    import Pie from "./charts/custom/Pie.vue";
-    import KPI from "./charts/custom/KPI.vue";
+    import Table from "./Table.vue";
+    import Pie from "./Pie.vue";
+    import KPI from "./KPI.vue";
     import {onMounted, ref} from "vue";
 
     const TYPES = {
