+++ ui/src/components/executions/Executions.vue
@@ -58,7 +58,7 @@
             </template>
 
             <template v-if="showStatChart()" #top>
-                <ChartsSection
+                <Sections
                     :charts="charts"
                     :show-default="true"
                     :full-size="true"
@@ -410,7 +410,7 @@
     import RunFast from "vue-material-design-icons/RunFast.vue";
     import ExecutionFilterLanguage from "../../composables/monaco/languages/filters/impl/executionFilterLanguage.ts";
     import FlowExecutionFilterLanguage from "../../composables/monaco/languages/filters/impl/flowExecutionFilterLanguage.js";
-    import ChartsSection from "../dashboard/components/ChartsSection.vue";
+    import Sections from "../dashboard/sections/Sections.vue";
 </script>
 
 <script>
